Move enum and constants to the root source folder
