+++ nw/constants.py
@@ -26,9 +26,7 @@ along with this program. If not, see <https://www.gnu.org/licenses/>.
 
 from PyQt5.QtCore import QCoreApplication, QT_TRANSLATE_NOOP
 
-from nw.constants.enum import (
-    nwItemClass, nwItemLayout, nwItemType, nwOutline
-)
+from nw.enum import nwItemClass, nwItemLayout, nwItemType, nwOutline
 
 def trConst(tString):
     """Wrapper function for locally translating constants.
