--- nw/constants/__init__.py
@@ -1,30 +1,0 @@
-# -*- coding: utf-8 -*-
-from nw.constants.constants import (
-    trConst, nwConst, nwLists, nwRegEx, nwFiles, nwKeyWords, nwLabels,
-    nwQuotes, nwUnicode, nwHtmlUnicode
-)
-from nw.constants.enum import (
-    nwAlert, nwDocAction, nwItemClass, nwItemLayout, nwItemType, nwOutline,
-    nwDocInsert, nwWidget
-)
-
-__all__ = [
-    "trConst",
-    "nwConst",
-    "nwLists",
-    "nwRegEx",
-    "nwFiles",
-    "nwKeyWords",
-    "nwLabels",
-    "nwQuotes",
-    "nwUnicode",
-    "nwHtmlUnicode",
-    "nwAlert",
-    "nwDocAction",
-    "nwItemClass",
-    "nwItemLayout",
-    "nwItemType",
-    "nwOutline",
-    "nwDocInsert",
-    "nwWidget",
-]
