+++ nw/core/tomd.py
@@ -26,8 +26,8 @@ along with this program. If not, see <https://www.gnu.org/licenses/>.
 
 import logging
 
-from nw.core.tokenizer import Tokenizer
 from nw.constants import nwLabels
+from nw.core.tokenizer import Tokenizer
 
 logger = logging.getLogger(__name__)
