+++ nw/enum.py
